Moved tests into integration and unit
