+++ tests/mindtrace/jobs/integration/test_orchestrator.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock
 from mindtrace.jobs.orchestrator import Orchestrator
 from mindtrace.jobs.local.client import LocalClient
 from mindtrace.jobs.redis.client import RedisClient
-from .conftest import create_test_job, unique_queue_name
+from ..conftest import create_test_job, unique_queue_name
 
 
 class TestOrchestrator:
@@ -125,22 +125,18 @@ class TestOrchestrator:
         queue_name = unique_queue_name("delete_test")
         client.declare_queue(queue_name)
         
-        # Verify queue exists
         assert queue_name in client.queues
         
-        # Delete queue through orchestrator
         orchestrator.delete_queue(queue_name)
         
-        # Verify queue is deleted
         assert queue_name not in client.queues
 
     def test_orchestrator_create_consumer_backend_rabbitmq(self):
         """Test creating RabbitMQ consumer backend - covers lines 107-110."""
         from mindtrace.jobs.rabbitmq.client import RabbitMQClient
         from mindtrace.jobs.types.job_specs import JobSchema
-        from .conftest import SampleJobInput, SampleJobOutput
+        from ..conftest import SampleJobInput, SampleJobOutput
         
-        # Create a mock class that inherits the name properly
         class MockRabbitMQClient:
             pass
         
@@ -148,17 +144,14 @@ class TestOrchestrator:
         rabbitmq_client = MockRabbitMQClient()
         orchestrator = Orchestrator(rabbitmq_client)
         
-        # Create a test schema
         test_schema = JobSchema(
             name="test_rabbitmq_job",
             input=SampleJobInput(),
             output=SampleJobOutput()
         )
         
-        # This should create a RabbitMQConsumerBackend (lines 107-110)
         consumer_backend = orchestrator.create_consumer_backend_for_schema(test_schema)
         
-        # Verify it's the correct type
         from mindtrace.jobs.rabbitmq.consumer_backend import RabbitMQConsumerBackend
         assert isinstance(consumer_backend, RabbitMQConsumerBackend)
         assert consumer_backend.queue_name == "test_rabbitmq_job"
@@ -166,9 +159,8 @@ class TestOrchestrator:
     def test_orchestrator_create_consumer_backend_unknown_type(self):
         """Test creating consumer backend with unknown backend type."""
         from mindtrace.jobs.types.job_specs import JobSchema
-        from .conftest import SampleJobInput, SampleJobOutput
+        from ..conftest import SampleJobInput, SampleJobOutput
         
-        # Create a mock backend with unknown type
         unknown_backend = MagicMock()
         unknown_backend.__class__.__name__ = "UnknownBackend"
         
@@ -180,6 +172,5 @@ class TestOrchestrator:
             output=SampleJobOutput()
         )
         
-        # Should raise ValueError for unknown backend type
         with pytest.raises(ValueError, match="Unknown backend type: UnknownBackend"):
             orchestrator.create_consumer_backend_for_schema(test_schema)
