+++ tests/mindtrace/jobs/integration/test_redis_backend.py
@@ -1,7 +1,7 @@
 import pytest
 import time
 from mindtrace.jobs.redis.client import RedisClient
-from .conftest import create_test_job, unique_queue_name
+from ..conftest import create_test_job, unique_queue_name
 
 
 @pytest.mark.redis
@@ -19,7 +19,6 @@ class TestRedisClient:
         result2 = self.client.declare_queue(queue_name)
         assert result2["status"] == "success"
         
-        # Cleanup
         self.client.delete_queue(queue_name)
     
     def test_publish_and_receive(self):
@@ -44,7 +43,6 @@ class TestRedisClient:
         count = self.client.count_queue_messages(queue_name)
         assert count == 0
         
-        # Cleanup
         self.client.delete_queue(queue_name)
     
     def test_priority_queue(self):
@@ -52,27 +50,22 @@ class TestRedisClient:
         queue_name = f"priority_queue_{int(time.time())}"
         self.client.declare_queue(queue_name, queue_type="priority")
         
-        # Create jobs with different priorities
         low_priority_job = create_test_job("low_priority_job", "low_priority_schema")
         high_priority_job = create_test_job("high_priority_job", "high_priority_schema")
         
-        # Publish lower priority first, then higher priority
         self.client.publish(queue_name, low_priority_job, priority=1)
         self.client.publish(queue_name, high_priority_job, priority=10)
         
-        # First received should be high priority
         first_received = self.client.receive_message(queue_name)
         assert first_received is not None
         assert isinstance(first_received, dict)
         assert first_received["schema_name"] == "high_priority_schema"
         
-        # Second received should be low priority
         second_received = self.client.receive_message(queue_name)
         assert second_received is not None
         assert isinstance(second_received, dict)
         assert second_received["schema_name"] == "low_priority_schema"
         
-        # Cleanup
         self.client.delete_queue(queue_name)
     
     def test_clean_and_delete(self):
@@ -108,7 +101,6 @@ class TestRedisClient:
         assert result["status"] == "success"
         assert "stack" in result["message"]
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_declare_priority_queue(self):
@@ -119,7 +111,6 @@ class TestRedisClient:
         assert result["status"] == "success"
         assert "priority" in result["message"]
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_declare_unknown_queue_type(self):
@@ -146,9 +137,7 @@ class TestRedisClient:
         queue_name = f"test_queue_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Create a job and remove job_id to test the generation
         test_job = create_test_job()
-        # Use model_dump and model_validate to remove job_id
         job_dict = test_job.model_dump()
         if 'job_id' in job_dict:
             del job_dict['job_id']
@@ -158,11 +147,9 @@ class TestRedisClient:
         assert isinstance(job_id, str)
         assert len(job_id) > 0
         
-        # Verify message was published
         count = self.client.count_queue_messages(queue_name)
         assert count == 1
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_receive_from_nonexistent_queue(self):
@@ -178,7 +165,6 @@ class TestRedisClient:
         result = self.client.receive_message(queue_name)
         assert result is None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_count_nonexistent_queue(self):
@@ -202,15 +188,12 @@ class TestRedisClient:
         queue_name = f"json_error_queue_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Directly push invalid JSON to the underlying queue
         instance = self.client.queues[queue_name]
         instance.push("invalid_json_data")
         
-        # Should return None due to JSON decode error
         result = self.client.receive_message(queue_name)
         assert result is None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_stack_queue_operations(self):
@@ -218,14 +201,12 @@ class TestRedisClient:
         queue_name = f"stack_test_{int(time.time())}"
         self.client.declare_queue(queue_name, queue_type="stack")
         
-        # Publish multiple jobs
         job1 = create_test_job("first_job", "first_schema")
         job2 = create_test_job("second_job", "second_schema")
         
         self.client.publish(queue_name, job1)
         self.client.publish(queue_name, job2)
         
-        # Stack should return last-in-first-out
         first_received = self.client.receive_message(queue_name)
         assert first_received is not None
         assert first_received["schema_name"] == "second_schema"
@@ -234,7 +215,6 @@ class TestRedisClient:
         assert second_received is not None
         assert second_received["schema_name"] == "first_schema"
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_lock_acquisition_failure_declare(self):
@@ -242,7 +222,6 @@ class TestRedisClient:
         import unittest.mock
         queue_name = f"lock_test_{int(time.time())}"
         
-        # Mock the redis lock to fail acquisition
         with unittest.mock.patch.object(self.client.redis, 'lock') as mock_lock:
             mock_lock_instance = unittest.mock.MagicMock()
             mock_lock_instance.acquire.return_value = False
@@ -256,10 +235,8 @@ class TestRedisClient:
         import unittest.mock
         queue_name = f"lock_test_{int(time.time())}"
         
-        # First declare the queue normally
         self.client.declare_queue(queue_name)
         
-        # Mock the redis lock to fail acquisition for delete
         with unittest.mock.patch.object(self.client.redis, 'lock') as mock_lock:
             mock_lock_instance = unittest.mock.MagicMock()
             mock_lock_instance.acquire.return_value = False
@@ -268,7 +245,6 @@ class TestRedisClient:
             with pytest.raises(BlockingIOError, match="Could not acquire distributed lock"):
                 self.client.delete_queue(queue_name)
         
-        # Cleanup normally
         self.client.delete_queue(queue_name)
 
     def test_lock_acquisition_failure_clean(self):
@@ -276,10 +252,8 @@ class TestRedisClient:
         import unittest.mock
         queue_name = f"lock_test_{int(time.time())}"
         
-        # First declare the queue normally
         self.client.declare_queue(queue_name)
         
-        # Mock the redis lock to fail acquisition for clean
         with unittest.mock.patch.object(self.client.redis, 'lock') as mock_lock:
             mock_lock_instance = unittest.mock.MagicMock()
             mock_lock_instance.acquire.return_value = False
@@ -288,7 +262,6 @@ class TestRedisClient:
             with pytest.raises(BlockingIOError, match="Could not acquire distributed lock"):
                 self.client.clean_queue(queue_name)
         
-        # Cleanup normally
         self.client.delete_queue(queue_name)
 
     def test_receive_message_unsupported_queue_type(self):
@@ -297,7 +270,6 @@ class TestRedisClient:
         queue_name = f"unsupported_queue_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Mock the queue instance to not have get or pop methods
         mock_instance = unittest.mock.MagicMock()
         del mock_instance.get
         del mock_instance.pop
@@ -305,11 +277,9 @@ class TestRedisClient:
         with self.client._local_lock:
             self.client.queues[queue_name] = mock_instance
         
-        # Should raise exception for unsupported queue type
         result = self.client.receive_message(queue_name)
         assert result is None  # Exception is caught and returns None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_clean_queue_exception_handling(self):
@@ -318,30 +288,23 @@ class TestRedisClient:
         queue_name = f"exception_queue_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Mock Redis operations to raise an exception
         with unittest.mock.patch.object(self.client.redis, 'llen', side_effect=Exception("Redis error")):
             with pytest.raises(Exception, match="Redis error"):
                 self.client.clean_queue(queue_name)
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_event_listener_unknown_queue_type(self):
         """Test event listener with unknown queue type - covers line 109 (continue)."""
         import json
         
-        # Create a fake event with unknown queue type
         event_data = json.dumps({"event": "declare", "queue": "test_queue", "queue_type": "unknown"})
         
-        # Publish to the events channel
         self.client.redis.publish(self.client.EVENTS_CHANNEL, event_data)
         
-        # Give some time for event processing
         import time
         time.sleep(0.1)
         
-        # The unknown queue type should be ignored (continue statement)
-        # Verify the queue wasn't added to local cache
         assert "test_queue" not in self.client.queues
 
     def test_event_listener_delete_event(self):
@@ -349,26 +312,19 @@ class TestRedisClient:
         import json
         import time
         
-        # First declare a queue normally
         queue_name = f"delete_event_test_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Verify queue exists in local cache
         assert queue_name in self.client.queues
         
-        # Send a delete event
         event_data = json.dumps({"event": "delete", "queue": queue_name})
         self.client.redis.publish(self.client.EVENTS_CHANNEL, event_data)
         
-        # Give some time for event processing
         time.sleep(0.1)
         
-        # Event listener should have removed the queue from local cache
-        # But we still need to clean up from Redis metadata
         if queue_name in self.client.queues:
             self.client.delete_queue(queue_name)
         else:
-            # Clean up metadata directly if queue was removed by event listener
             try:
                 self.client.redis.hdel(self.client.METADATA_KEY, queue_name)
             except:
@@ -383,12 +339,10 @@ class TestRedisClient:
         test_job = create_test_job()
         job_id = self.client.publish(queue_name, test_job, priority=5)
         
-        # Receive the message - should use 'get' method
         received = self.client.receive_message(queue_name)
         assert received is not None
         assert received["id"] == test_job.id
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_receive_message_missing_both_methods(self):
@@ -397,9 +351,7 @@ class TestRedisClient:
         queue_name = f"no_methods_test_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Mock the queue instance to have neither get nor pop
         mock_instance = unittest.mock.MagicMock()
-        # Remove both get and pop attributes
         if hasattr(mock_instance, 'get'):
             delattr(mock_instance, 'get')
         if hasattr(mock_instance, 'pop'):
@@ -408,11 +360,9 @@ class TestRedisClient:
         with self.client._local_lock:
             self.client.queues[queue_name] = mock_instance
         
-        # Should return None due to exception handling
         result = self.client.receive_message(queue_name)
         assert result is None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_stack_queue_event_listener(self):
@@ -420,18 +370,13 @@ class TestRedisClient:
         import json
         import time
         
-        # Send stack queue declare event
         queue_name = f"stack_event_{int(time.time())}"
         event_data = json.dumps({"event": "declare", "queue": queue_name, "queue_type": "stack"})
         self.client.redis.publish(self.client.EVENTS_CHANNEL, event_data)
         
-        # Give time for event processing
         time.sleep(0.1)
         
-        # The queue should have been added to local cache
-        # This test helps cover the stack queue creation path in event listener
         try:
-            # Clean up metadata
             self.client.redis.hdel(self.client.METADATA_KEY, queue_name)
         except:
             pass
@@ -441,18 +386,13 @@ class TestRedisClient:
         import json
         import time
         
-        # Send priority queue declare event  
         queue_name = f"priority_event_{int(time.time())}"
         event_data = json.dumps({"event": "declare", "queue": queue_name, "queue_type": "priority"})
         self.client.redis.publish(self.client.EVENTS_CHANNEL, event_data)
         
-        # Give time for event processing
         time.sleep(0.1)
         
-        # The queue should have been added to local cache
-        # This test helps cover the priority queue creation path in event listener
         try:
-            # Clean up metadata
             self.client.redis.hdel(self.client.METADATA_KEY, queue_name)
         except:
             pass
@@ -461,18 +401,15 @@ class TestRedisClient:
         """Test receive_message method paths - covers lines 214-215."""
         import unittest.mock
         
-        # Test the elif branch with pop method
         queue_name = f"pop_test_{int(time.time())}"
         self.client.declare_queue(queue_name, queue_type="stack")  # Stack uses pop
         
         test_job = create_test_job()
         self.client.publish(queue_name, test_job)
         
-        # This should trigger the pop method path
         received = self.client.receive_message(queue_name)
         assert received is not None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_receive_message_no_get_or_pop(self):
@@ -482,17 +419,14 @@ class TestRedisClient:
         queue_name = f"no_attrs_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Create a mock that doesn't have get or pop attributes at all
         mock_instance = type('MockQueue', (), {})()
         
         with self.client._local_lock:
             self.client.queues[queue_name] = mock_instance
         
-        # This should trigger the Exception on line 228
         result = self.client.receive_message(queue_name)
         assert result is None  # Exception caught and returns None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_event_listener_exception_handling(self):
@@ -500,22 +434,17 @@ class TestRedisClient:
         import json
         import unittest.mock
         
-        # Send malformed JSON to trigger exception in event processing
         self.client.redis.publish(self.client.EVENTS_CHANNEL, "{malformed")
         
-        # Send invalid event type to trigger exception
         event_data = json.dumps({"event": "invalid", "queue": "test"})
         self.client.redis.publish(self.client.EVENTS_CHANNEL, event_data)
         
-        # Send event missing required fields
         event_data = json.dumps({"event": "declare"})  # Missing queue name
         self.client.redis.publish(self.client.EVENTS_CHANNEL, event_data)
         
-        # Give time for event processing
         import time
         time.sleep(0.1)
         
-        # All exceptions should be caught and passed silently
 
     def test_publish_exception_handling(self):
         """Test publish exception handling - covers lines 214-215."""
@@ -523,25 +452,20 @@ class TestRedisClient:
         queue_name = f"exception_test_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Create a mock instance that raises on push
         mock_instance = unittest.mock.MagicMock()
         mock_instance.push.side_effect = Exception("Push failed")
         
         with self.client._local_lock:
             self.client.queues[queue_name] = mock_instance
         
-        # Create a test job that will work with model_dump
         test_job = unittest.mock.MagicMock()
         test_job.model_dump.return_value = {"id": "test", "data": "test"}
         
-        # Should re-raise the exception
         with pytest.raises(Exception, match="Push failed"):
             self.client.publish(queue_name, test_job)
         
-        # Verify the mock was called correctly
         mock_instance.push.assert_called_once()
         
-        # Cleanup
         self.client.delete_queue(queue_name)
 
     def test_receive_message_get_method(self):
@@ -550,7 +474,6 @@ class TestRedisClient:
         queue_name = f"get_test_{int(time.time())}"
         self.client.declare_queue(queue_name)
         
-        # Create a mock instance that has a get method that raises an exception
         class TestQueue:
             def get(self, block=False, timeout=None):
                 raise Exception("Test exception from get")
@@ -560,13 +483,10 @@ class TestRedisClient:
         
         mock_instance = TestQueue()
         
-        # Replace the queue instance with our test class
         with self.client._local_lock:
             self.client.queues[queue_name] = mock_instance
         
-        # The exception from get() should be caught and converted to None
         result = self.client.receive_message(queue_name)
         assert result is None
         
-        # Cleanup
         self.client.delete_queue(queue_name)
